:recycle: getForms
