+++ src/client/form/forms.rs
@@ -1,49 +1,14 @@
 use crate::query::query_string;
-use anyhow::anyhow;
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
+use crate::client::HEADER_KEY;
+use crate::response::error::{ErrorResponse, ErrorResponseBody};
+use crate::response::form::forms::{GetFormsResponse, GetFormsResponseBody};
 
 const RESOURCE: &str = "forms";
-const HEADER_KEY: &str = "X-Collaboflow-Authorization";
 
 pub struct Forms {
     url: String,
     authorization_header: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct GetFormsResponse {
-    pub app_cd: i32,
-    pub offset: i32,
-    pub limit: i32,
-    pub total_count: i32,
-    pub error: bool,
-    pub records: Vec<FormRecord>,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct FormRecord {
-    pub id: i32,
-    pub name: String,
-    pub category: FormCategory,
-    pub description: String,
-    pub status: String,
-    pub r#type: String,
-    pub version: i32,
-    pub create_date: String,
-    pub start_date: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct FormCategory {
-    pub id: i32,
-    pub name: String,
-}
-
-#[derive(Debug, Deserialize, Clone, Serialize)]
-pub struct GetFormsErrorResponse {
-    pub error: bool,
-    pub messages: Vec<String>,
 }
 
 impl Forms {
@@ -57,19 +22,57 @@ impl Forms {
     pub async fn get(
         &self,
         query_params: HashMap<String, String>,
-    ) -> anyhow::Result<GetFormsResponse> {
+    ) -> Result<GetFormsResponse, ErrorResponse> {
         let request_url = format!("{}?{}", &self.url, query_string(query_params));
 
         let http_client = reqwest::Client::new();
-        let resp = http_client
+        let result = http_client
             .get(request_url)
             .header(HEADER_KEY, &self.authorization_header)
             .send()
-            .await?;
+            .await;
 
-        match resp.json::<GetFormsResponse>().await {
-            Ok(forms) => Ok(forms),
-            Err(err) => Err(anyhow!(err)),
+        match result {
+            Ok(resp) => {
+                let status = resp.status().as_u16();
+
+                if status == 200 {
+                    match resp.json::<GetFormsResponseBody>().await {
+                        Ok(body) => Ok(GetFormsResponse { status, body }),
+                        Err(err) => {
+                            let body = ErrorResponseBody {
+                                error: true,
+                                messages: vec![err.to_string()],
+                            };
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                    }
+                } else {
+                    match resp.json::<ErrorResponseBody>().await {
+                        Ok(body) => {
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                        Err(err) => {
+                            let body = ErrorResponseBody {
+                                error: true,
+                                messages: vec![err.to_string()],
+                            };
+                            let error_response = ErrorResponse { status, body };
+                            Err(error_response)
+                        }
+                    }
+                }
+            }
+            Err(err) => {
+                let body = ErrorResponseBody {
+                    error: true,
+                    messages: vec![err.to_string()],
+                };
+                let error_response = ErrorResponse { status: 500, body };
+                Err(error_response)
+            }
         }
     }
 }
