+++ src/response.rs
@@ -58,3 +58,9 @@ pub struct ProcessFrom {
     pub name: String,
     pub r#type: String,
 }
+
+#[derive(Debug, Deserialize, Clone, Serialize)]
+pub struct FormCategory {
+    pub id: i32,
+    pub name: String,
+}
